# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Grocery Multi-Vendor API",
        "description": "API documentation for Grocery Multi-Vendor Platform. This API supports Admin, Vendor, Delivery Boy, and Customer roles.",
        "contact": {
            "email": "support@grocery.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/vendors": {
            "get": {
                "tags": [
                    "Vendors"
                ],
                "summary": "List all vendors (Public)",
                "description": "Display a listing of vendors",
                "operationId": "3a74345b7d6981df67075b76d655d93e",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (active, inactive, pending)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "verified",
                        "in": "query",
                        "description": "Filter by verification status (1 or 0)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search vendors by store name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vendors retrieved successfully"
                    }
                }
            }
        },
        "/api/v1/vendors/{id}": {
            "get": {
                "tags": [
                    "Vendors"
                ],
                "summary": "Get vendor details (Public)",
                "description": "Display the specified vendor",
                "operationId": "550d6f5c6b07df747f124874d860f603",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vendor retrieved successfully"
                    },
                    "404": {
                        "description": "Vendor not found"
                    }
                }
            }
        },
        "/api/v1/vendors/{id}/approve": {
            "post": {
                "tags": [
                    "Vendors"
                ],
                "summary": "Approve vendor (Admin only)",
                "description": "Approve vendor (Admin only)",
                "operationId": "79cb620c6035fa46ff23e48d8d0003a5",
                "responses": {
                    "200": {
                        "description": "Vendor approved successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/vendors/{id}/reject": {
            "post": {
                "tags": [
                    "Vendors"
                ],
                "summary": "Reject vendor (Admin only)",
                "description": "Reject vendor (Admin only)",
                "operationId": "328035e46207f845fbf8df7422aedc5e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rejection_reason"
                                ],
                                "properties": {
                                    "rejection_reason": {
                                        "type": "string",
                                        "example": "Incomplete documentation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vendor rejected successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}